pysrc/requirements.txt
inotify-simple
rpi-libcamera
numpy
simplejpeg
